Revert "refactor: move ChooseCheckGroup to page folder (#1021)"
This reverts commit cf6789171f353a4839b03486decb5d3f6d7fec02.
