+++ src/components/CheckEditor/FormComponents/CheckStatusInfo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { GrafanaTheme2 } from '@grafana/data';
-import { Icon, Stack, TextLink, useStyles2 } from '@grafana/ui';
+import { Badge, Icon, Stack, TextLink, useStyles2 } from '@grafana/ui';
 import { css } from '@emotion/css';
 
 import { CheckStatus } from 'types';
@@ -38,6 +38,14 @@ export const CheckStatusInfo = ({ description, docsLink }: CheckStatusInfoProps)
       </Toggletip>
     </Stack>
   );
+};
+
+export const NewStatusBadge = ({ status, className }: { status: CheckStatus; className?: string }) => {
+  if (![CheckStatus.EXPERIMENTAL, CheckStatus.PRIVATE_PREVIEW, CheckStatus.PUBLIC_PREVIEW].includes(status)) {
+    return null;
+  }
+
+  return <Badge text={'NEW'} color={'orange'} className={className} />;
 };
 
 const getStyles = (theme: GrafanaTheme2) => ({
