--- src/components/CheckStatusInfo/index.ts
@@ -1,1 +1,0 @@
-export { CheckStatusInfo, type CheckStatusInfoProps } from './CheckStatusInfo';
