--- src/components/NewStatusBadge/index.ts
@@ -1,1 +1,0 @@
-export { NewStatusBadge } from './NewStatusBadge';
